Make api exceptions and permissions common modules
